+++ apps/backend/src/modules/routes/libs/types/types.ts
@@ -1,5 +1,6 @@
 export { type UserRoute } from "./user-route.js";
 export {
+	type FileUploadResponseDto,
 	type RouteConstructRequestDto,
 	type RouteCreateRequestDto,
 	type RouteFindAllOptions,
@@ -7,5 +8,6 @@ export {
 	type RouteGetByIdResponseDto,
 	type RoutePatchRequestDto,
 	type RoutePatchResponseDto,
+	type RouteUploadImageResponseDto,
 	type UserRouteStatusType,
 } from "@smartscapes/shared";