+++ apps/backend/src/modules/routes/route.entity.ts
@@ -1,6 +1,9 @@
 import { type Entity, type LineStringGeometry } from "~/libs/types/types.js";
 
-import { type UserRoute } from "./libs/types/types.js";
+import {
+	type RouteUploadImageResponseDto,
+	type UserRoute,
+} from "./libs/types/types.js";
 
 class RouteEntity implements Entity {
 	private createdByUserId: number;
@@ -14,6 +17,8 @@ class RouteEntity implements Entity {
 	private geometry: LineStringGeometry;
 
 	private id: null | number;
+
+	private images: RouteUploadImageResponseDto[];
 
 	private name: string;
 
@@ -32,6 +37,7 @@ class RouteEntity implements Entity {
 		duration,
 		geometry,
 		id,
+		images,
 		name,
 		pois,
 		userRoute,
@@ -42,6 +48,7 @@ class RouteEntity implements Entity {
 		duration: number;
 		geometry: LineStringGeometry;
 		id: null | number;
+		images: RouteUploadImageResponseDto[];
 		name: string;
 		pois: {
 			id: number;
@@ -59,6 +66,7 @@ class RouteEntity implements Entity {
 		this.pois = pois;
 		this.createdByUserId = createdByUserId;
 		this.userRoute = userRoute?.[0] ?? null;
+		this.images = images;
 	}
 
 	public static initialize(data: {
@@ -68,6 +76,7 @@ class RouteEntity implements Entity {
 		duration: number;
 		geometry: LineStringGeometry;
 		id: number;
+		images: RouteUploadImageResponseDto[];
 		name: string;
 		pois: {
 			id: number;
@@ -83,6 +92,7 @@ class RouteEntity implements Entity {
 			duration: data.duration,
 			geometry: data.geometry,
 			id: data.id,
+			images: data.images,
 			name: data.name,
 			pois: data.pois,
 			userRoute: data.userRoute ?? null,
@@ -95,6 +105,7 @@ class RouteEntity implements Entity {
 		duration,
 		geometry,
 		id,
+		images,
 		name,
 		pois,
 	}: {
@@ -103,6 +114,7 @@ class RouteEntity implements Entity {
 		duration: number;
 		geometry: LineStringGeometry;
 		id: number;
+		images: RouteUploadImageResponseDto[];
 		name: string;
 		pois: {
 			id: number;
@@ -117,6 +129,7 @@ class RouteEntity implements Entity {
 			duration,
 			geometry,
 			id,
+			images,
 			name,
 			pois,
 			userRoute: null,
@@ -152,6 +165,7 @@ class RouteEntity implements Entity {
 			duration,
 			geometry,
 			id: null,
+			images: [],
 			name,
 			pois,
 			userRoute: userRoute ?? null,
@@ -165,6 +179,7 @@ class RouteEntity implements Entity {
 		duration: number;
 		geometry: LineStringGeometry;
 		id: number;
+		images: RouteUploadImageResponseDto[];
 		name: string;
 		pois: {
 			id: number;
@@ -180,6 +195,7 @@ class RouteEntity implements Entity {
 			duration: this.duration,
 			geometry: this.geometry,
 			id: this.id as number,
+			images: this.images,
 			name: this.name,
 			pois: this.pois as {
 				id: number;
@@ -196,6 +212,7 @@ class RouteEntity implements Entity {
 		duration: number;
 		geometry: LineStringGeometry;
 		id: number;
+		images: RouteUploadImageResponseDto[];
 		name: string;
 		pois: {
 			id: number;
@@ -209,6 +226,7 @@ class RouteEntity implements Entity {
 			duration: this.duration,
 			geometry: this.geometry,
 			id: this.id as number,
+			images: this.images,
 			name: this.name,
 			pois: this.pois as {
 				id: number;
@@ -248,6 +266,7 @@ class RouteEntity implements Entity {
 		duration: number;
 		geometry: LineStringGeometry;
 		id: number;
+		images: RouteUploadImageResponseDto[];
 		name: string;
 		pois: {
 			id: number;
@@ -262,6 +281,7 @@ class RouteEntity implements Entity {
 			duration: this.duration,
 			geometry: this.geometry,
 			id: this.id as number,
+			images: this.images,
 			name: this.name,
 			pois: this.pois as Required<
 				Pick<RouteEntity["pois"][number], "id" | "name" | "visitOrder">