+++ apps/backend/src/modules/user-routes/libs/types/type.ts
@@ -1,8 +1,7 @@
 export {
-	type UserRouteCreateRequestDto,
 	type UserRouteFinishRequestDto,
 	type UserRouteParameters,
+	type UserRouteQueryRequestDto,
 	type UserRouteResponseDto,
-	type UserRouteStartRequestDto,
 	type UserRouteStatusType,
 } from "@smartscapes/shared";