+++ apps/backend/src/modules/user-routes/libs/validation-schemas/validation-schemas.ts
@@ -1,6 +1,5 @@
 export {
-	userRouteCreateValidationSchema,
 	userRouteDeleteValidationSchema,
 	userRouteFinishValidationSchema,
-	userRouteStartValidationSchema,
+	userRouteQueryValidationSchema,
 } from "@smartscapes/shared";