+++ apps/frontend/src/pages/route-details/route-details.tsx
@@ -1,11 +1,9 @@
-import image1 from "~/assets/images/route-details/placeholder-image-1.png";
-import image2 from "~/assets/images/route-details/placeholder-image-2.png";
-import image3 from "~/assets/images/route-details/placeholder-image-3.png";
 import {
 	Button,
 	FeatureGallery,
 	Input,
 	Loader,
+	MapProvider,
 	TextArea,
 } from "~/libs/components/components.js";
 import { DataStatus, PermissionKey } from "~/libs/enums/enums.js";
@@ -17,6 +15,7 @@ import {
 	useCallback,
 	useEffect,
 	useParams,
+	useRef,
 	useState,
 } from "~/libs/hooks/hooks.js";
 import { type ReviewRequestDto } from "~/modules/reviews/reviews.js";
@@ -61,6 +60,27 @@ const RouteDetails = (): React.JSX.Element => {
 		user &&
 			checkHasPermission([PermissionKey.MANAGE_ROUTES], user.group.permissions),
 	);
+	const fileInputReference = useRef<HTMLInputElement | null>(null);
+
+	const handleFileUpload = useCallback(
+		(event: React.ChangeEvent<HTMLInputElement>) => {
+			const file = event.target.files?.[0];
+
+			if (!file) {
+				return;
+			}
+
+			void dispatch(
+				routeActions.uploadImage({ file, id: route?.id as number }),
+			);
+			event.target.value = "";
+		},
+		[dispatch, route],
+	);
+
+	const handleTriggerFileUpload = useCallback(() => {
+		fileInputReference.current?.click();
+	}, []);
 
 	const handleToggleEditMode = useCallback(() => {
 		setIsEditMode((isEditMode) => !isEditMode);
@@ -107,9 +127,20 @@ const RouteDetails = (): React.JSX.Element => {
 		}
 	}, [route?.userRoute?.id, dispatch]);
 
+	const handleDeleteImage = useCallback(
+		(id: number) => {
+			void dispatch(routeActions.deleteImage(id));
+		},
+		[dispatch],
+	);
+
 	useEffect(() => {
 		void dispatch(routeActions.getById(Number(routeId)));
 	}, [dispatch, routeId]);
+
+	useEffect(() => {
+		handleResetFormValues();
+	}, [handleResetFormValues]);
 
 	useEffect(() => {
 		void dispatch(routeActions.getReviews({ routeId: Number(routeId) }));
@@ -138,7 +169,7 @@ const RouteDetails = (): React.JSX.Element => {
 		return <></>;
 	}
 
-	const { description, id, name, pois } = route;
+	const { description, id, images, name, pois } = route;
 
 	const hasDescription = Boolean(description);
 
@@ -161,17 +192,17 @@ const RouteDetails = (): React.JSX.Element => {
 				) : (
 					<>
 						<h1 className={styles["label"]}>{name}</h1>
-						{isAuthorized && (
-							<div className={styles["controls-container"]}>
-								{hasEditPermissions && (
-									<div className={styles["edit-button-container"]}>
-										<Button
-											label="Edit"
-											onClick={handleToggleEditMode}
-											variant="outlined"
-										/>
-									</div>
-								)}
+						<div className={styles["controls-container"]}>
+							{hasEditPermissions && (
+								<div className={styles["edit-button-container"]}>
+									<Button
+										label="Edit"
+										onClick={handleToggleEditMode}
+										variant="outlined"
+									/>
+								</div>
+							)}
+							{isAuthorized && (
 								<div className={styles["save-button-container"]}>
 									<Button
 										disabled={isSaving}
@@ -183,25 +214,55 @@ const RouteDetails = (): React.JSX.Element => {
 										pressed={isSaved}
 									/>
 								</div>
-							</div>
-						)}
+							)}
+						</div>
 					</>
 				)}
 			</div>
 			<FeatureGallery
 				slides={[
-					{ content: image1, type: "image" },
-					{ content: image2, type: "image" },
-					{ content: image3, type: "image" },
+					{
+						content: <MapProvider />,
+					},
+					...images.map((image) => ({
+						content: (
+							<img
+								alt="point of interest"
+								className={styles["image"]}
+								src={image.url}
+							/>
+						),
+						...(isEditMode && {
+							onDelete: (): void => {
+								handleDeleteImage(image.id);
+							},
+						}),
+					})),
 				]}
 			/>
 			{isEditMode ? (
-				<TextArea
-					control={control}
-					errors={errors}
-					label="Description"
-					name="description"
-				/>
+				<>
+					<input
+						accept="image/*"
+						onChange={handleFileUpload}
+						ref={fileInputReference}
+						style={{ display: "none" }}
+						type="file"
+					/>
+					<div className={styles["upload-button"]}>
+						<Button
+							label="Upload image"
+							onClick={handleTriggerFileUpload}
+							variant="outlined"
+						/>
+					</div>
+					<TextArea
+						control={control}
+						errors={errors}
+						label="Description"
+						name="description"
+					/>
+				</>
 			) : (
 				hasDescription && <p className={styles["description"]}>{description}</p>
 			)}