+++ packages/shared/src/modules/routes/libs/types/route-get-all-item-response-dto.type.ts
@@ -1,4 +1,5 @@
 import { type LineStringGeometry } from "../../../../libs/types/types.js";
+import { type FileUploadResponseDto } from "../../../files/libs/types/types.js";
 
 type RouteGetAllItemResponseDto = {
 	createdByUserId: number;
@@ -6,6 +7,7 @@ type RouteGetAllItemResponseDto = {
 	duration: number;
 	geometry: LineStringGeometry;
 	id: number;
+	images: Pick<FileUploadResponseDto, "id" | "url">[];
 	name: string;
 	pois: {
 		id: number;