+++ packages/shared/src/modules/routes/libs/types/route-get-by-id-response-dto.type.ts
@@ -1,4 +1,5 @@
 import { type LineStringGeometry } from "../../../../libs/types/types.js";
+import { type FileUploadResponseDto } from "../../../files/libs/types/types.js";
 import { type UserRouteStatusType } from "../../../user-routes/user-routes.js";
 
 type RouteGetByIdResponseDto = {
@@ -8,6 +9,7 @@ type RouteGetByIdResponseDto = {
 	duration: number;
 	geometry: LineStringGeometry;
 	id: number;
+	images: Pick<FileUploadResponseDto, "id" | "url">[];
 	name: string;
 	pois: {
 		id: number;