+++ packages/shared/src/modules/user-routes/libs/enums/user-route-api-path.enum.ts
@@ -1,5 +1,6 @@
 const UserRouteApiPath = {
 	$ID: "/:id",
+	CREATE: "/create",
 	FINISH: "/finish",
 	ROOT: "/",
 	START: "/start",