+++ packages/shared/src/modules/user-routes/libs/types/types.ts
@@ -1,7 +1,6 @@
-export { type UserRouteCreateRequestDto } from "./user-route-create-request-dto.type.js";
 export { type UserRouteFinishRequestDto } from "./user-route-finish-request-dto.type.js";
 export { type UserRouteParameters } from "./user-route-parameters-request.type.js";
 export { type UserRoutePatchRequestDto } from "./user-route-patch-request-dto.type.js";
+export { type UserRouteQueryRequestDto } from "./user-route-query-request-dto.type.js";
 export { type UserRouteResponseDto } from "./user-route-response-dto.type.js";
-export { type UserRouteStartRequestDto } from "./user-route-start-request-dto.type.js";
 export { type UserRouteStatusType } from "./user-route-status.type.js";