+++ packages/shared/src/modules/user-routes/libs/types/user-route-start-request-dto.type.ts
@@ -1,5 +1,0 @@
-type UserRouteStartRequestDto = {
-	routeId: number;
-};
-
-export { type UserRouteStartRequestDto };