+++ packages/shared/src/modules/user-routes/user-routes.ts
@@ -1,18 +1,17 @@
 export { UserRouteApiPath, UserRouteStatus } from "./libs/enums/enums.js";
 export {
-	type UserRouteCreateRequestDto,
 	type UserRouteFinishRequestDto,
 	type UserRouteParameters,
 	type UserRoutePatchRequestDto,
+	type UserRouteQueryRequestDto,
 	type UserRouteResponseDto,
-	type UserRouteStartRequestDto,
 	type UserRouteStatusType,
 } from "./libs/types/types.js";
 export {
-	userRouteCreate as userRouteCreateValidationSchema,
 	userRouteDeleteParameters as userRouteDeleteValidationSchema,
 	userRouteFinish as userRouteFinishValidationSchema,
 	userRouteParameters as userRouteParametersValidationSchema,
 	userRoutePatch as userRoutePatchValidationSchema,
 	userRouteStart as userRouteStartValidationSchema,
 } from "./validation-schemas/validation-schemas.js";
+export { userRouteQuery as userRouteQueryValidationSchema } from "./validation-schemas/validation-schemas.js";